IAV SDK compiled against 3rd party OMID
