--- OMSDK-1.3.7/OMSDK_Pubnativenet.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
