--- OMSDK-1.3.7/OMSDK_Pubnativenet.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-module OMSDK_Pubnativenet {
-    header "Headers/OMIDImports.h"
-    export *
-}
