--- OMSDK-1.3.7/OMSDK_Pubnativenet.framework/Versions/Current
@@ -1,1 +1,0 @@
-A
